+++ R/getContent.R
@@ -22,6 +22,7 @@ getContent <-
              "datetime",
              "edittime",
              "press",
+             "reporter",
              "title",
              "body"
            )) {
@@ -42,6 +43,7 @@ news_content <-
              "datetime",
              "edittime",
              "press",
+             "reporter",
              "title",
              "body"
            )) {
@@ -63,6 +65,7 @@ news_content <-
       datetime <- "page is not news section."
       edittime <- "page is not news section."
       press <- "page is not news section."
+      reporter <- "page is not news section."
       body <- "page is not news section."
       section <- "page is not news section."
     } else {
@@ -71,6 +74,7 @@ news_content <-
       datetime <- getContentDatetime(html_obj)
       edittime <- getContentEditDatetime(html_obj)
       press <- getContentPress(html_obj)
+      reporter <- news_content_reporter(html_obj)
       body <- getContentBody(html_obj)
       section <- getSection(turl)
     }
@@ -84,6 +88,7 @@ news_content <-
       datetime = datetime,
       edittime = edittime,
       press = press,
+      reporter = reporter,
       title = title,
       body = body,
       section = section
@@ -138,6 +143,22 @@ getContentPress <-
       press <- rvest::html_attr(node, press_attr)
     }
     return(press[1])
+  }
+
+news_content_reporter <-
+  function(html_obj,
+           reporter_node_info = c(
+             ".media_end_head_journalist_name",
+             ".byline_s"
+           )) {
+    node <- rvest::html_nodes(html_obj, paste(reporter_node_info, collapse = ", "))
+    reporter <- trimws(rvest::html_text(node))
+    reporter <- reporter[nchar(reporter) > 0]
+
+    if (length(reporter) == 0) {
+      return(NA_character_)
+    }
+    return(reporter[1])
   }
 
 getContentBody <-